support scalar parameters in type inference
When node parameters are scalars (e.g., LIF(tau=np.float32(4.0))), the resulting input_type
contains tuples of length 0, since scalar parameters don't specify the layer size.
In these cases, we need to infer the input dimensions from the preceding nodes.
This is done by checking if the input_type is None or if any of the values are None or empty tuples.